Solved Problem
solved today problem

+++ 20200107/sumBetweenTwoInteger-kotlin-saj/src/main/kotlin/App.kt
@@ -1,7 +1,7 @@
 fun sumBetweenTwoInteger(a: Int, b: Int): Long {
     return when {
-        a < b -> (a..b).sum().toLong()
-        a > b -> (b..a).sum().toLong()
+        a < b -> (a.toLong()..b.toLong()).sum().toLong()
+        a > b -> (b.toLong()..a.toLong()).sum().toLong()
         else -> a.toLong()
     }
 }
